Small refactor to make it easier to register new handlers for the database interaction classes
